Refactor Hero attack method, and some cleaning

+++ src/Soldier.java
@@ -16,7 +16,7 @@ public class Soldier {
         return false;
     }
 
-    public void getDamage(int attackPower) {
+    public void getDamage(double attackPower) {
         this.setCurrentHealth(currentHealth - attackPower);             // did not check the skills!
     }
     //------------------------------------------- Getters && Setters
